+++ modules/services/views/src/main/java/au/com/shiftyjelly/pocketcasts/views/swipe/SwipeAction.kt
@@ -18,6 +18,7 @@ enum class SwipeAction(
     Archive("archive"),
     Unarchive("unarchive"),
     RemoveFromPlaylist("playlist_remove"),
+    DeleteUserEpisode("delete"),
     ;
 
     override fun contentDescription(context: Context) = when (this) {
@@ -28,6 +29,7 @@ enum class SwipeAction(
         Archive -> context.getString(LR.string.archive)
         Unarchive -> context.getString(LR.string.unarchive)
         RemoveFromPlaylist -> context.getString(LR.string.remove_from_playlist)
+        DeleteUserEpisode -> context.getString(LR.string.delete)
     }
 
     override fun backgroundTint(context: Context): Int {
@@ -39,6 +41,7 @@ enum class SwipeAction(
             Archive -> UR.attr.support_06
             Unarchive -> UR.attr.support_06
             RemoveFromPlaylist -> UR.attr.support_05
+            DeleteUserEpisode -> UR.attr.support_05
         }
         return context.getThemeColor(id)
     }
@@ -53,5 +56,6 @@ enum class SwipeAction(
         Archive -> IR.drawable.ic_archive
         Unarchive -> IR.drawable.ic_unarchive
         RemoveFromPlaylist -> IR.drawable.ic_delete
+        DeleteUserEpisode -> IR.drawable.ic_delete
     }
 }